--- app/src/main/java/com/example/gear_kotlin/di/TestService.kt
@@ -1,13 +1,0 @@
-package com.example.gear_kotlin.di
-
-import android.app.Service
-import android.content.Intent
-import android.os.IBinder
-
-class TestService {
-    var i: Int = 0
-
-    init {
-        i = 1
-    }
-}
