+++ app/src/main/java/com/example/gear_kotlin/di/module/ActivityProvider.kt
@@ -1,18 +1,17 @@
-package com.example.gear_kotlin.di
+package com.example.gear_kotlin.di.module
 
 import com.example.gear_kotlin.ui.MainActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
 @Module
-abstract class ActivitiesBuilder {
+abstract class ActivityProvider {
     /*
  * We modify our ActivityModule by adding the
  * FragmentModule to the Activity which contains
  * the fragment
  */
-    @ContributesAndroidInjector(modules = [FragmentProvider::class, TestModule::class])
+    @ContributesAndroidInjector(modules = [FragmentProvider::class])
     abstract fun contributeMainActivity(): MainActivity
-
 
 }
