+++ app/src/main/java/com/example/gear_kotlin/di/module/DBModule.kt
@@ -1,0 +1,19 @@
+package com.example.gear_kotlin.di.module
+
+import com.example.gear_kotlin.GearApplication
+import com.example.local.AppDatabase
+import com.example.local.dao.UserDao
+import com.example.repository.UserRepository
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module()
+class DBModule {
+
+    @Singleton
+    @Provides
+    fun provideUserDao(db: AppDatabase): UserDao {
+        return db.userDao()
+    }
+}
