+++ app/src/main/java/com/example/gear_kotlin/di/module/ViewModelModule.kt
@@ -1,30 +1,26 @@
-package com.example.gear_kotlin.di
+package com.example.gear_kotlin.di.module
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.example.gear_kotlin.di.DaggerViewModelFactory
+import com.example.gear_kotlin.di.ViewModelKey
 import com.example.gear_kotlin.viewmodel.ThirdFragmentViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
+import javax.inject.Singleton
 
 @Module
 abstract class ViewModelModule {
 
-    /*
-     * This method basically says
-     * inject this object into a Map using the @IntoMap annotation,
-     * with the  MovieListViewModel.class as key,
-     * and a Provider that will build a MovieListViewModel
-     * object.
-     *
-     * */
+    @Binds
+    abstract fun bindViewModelFactory(factory: DaggerViewModelFactory): ViewModelProvider.Factory
 
+    @Singleton
     @Binds
     @IntoMap
     @ViewModelKey(ThirdFragmentViewModel::class)
     abstract fun thirdFragmentViewModel(model: ThirdFragmentViewModel): ViewModel
 
 
-    @Binds
-    abstract fun bindViewModelFactory(factory: DaggerViewModelFactory): ViewModelProvider.Factory
 }
