+++ app/src/main/java/com/example/gear_kotlin/ui/MainActivity.kt
@@ -8,25 +8,34 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.example.gear_kotlin.R
-import com.example.gear_kotlin.di.TestService
+import com.example.gear_kotlin.viewmodel.ThirdFragmentViewModel
 
 import dagger.android.support.DaggerAppCompatActivity
 import kotlinx.android.synthetic.main.activity_main.*
 import javax.inject.Inject
 
 class MainActivity : DaggerAppCompatActivity() {
+    @Inject
+    lateinit var viewModel: ThirdFragmentViewModel
 
+    @Inject
+    lateinit var viewModel1: ThirdFragmentViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         setSupportActionBar(toolbar)
-        Log.i("hanhmh1203","value i ${service.i}")
     }
-//    @Inject
-//    internal lateinit var viewModelFactory: ViewModelProvider.Factory
-    @Inject
-    lateinit var service: TestService
+
+    override fun onResume() {
+        super.onResume()
+//        viewModel.increaseI()
+//        viewModel.increaseI()
+//        viewModel.increaseI()
+//
+//
+//        viewModel1.increaseI()
+    }
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         // Inflate the menu; this adds items to the action bar if it is present.
         menuInflater.inflate(R.menu.menu_main, menu)
